// apps/frontend/src/app/types/gym-filters.ts
export type gymFilters = {
  searchParamLimit?: string | null;
  searchParamSortDirection?: string | null;
  searchParamMinPrice?: string | null;
  searchParamMaxPrice?: string | null;
  searchParamLocation?: string | string[] | null;
  searchParamOptions?: string | string[] | null;
  searchParamIsOnlyVerified?: string | null;
}